feat. 세션 초기화 문제 해결 (시작하자마자 객체 생성)

+++ SECUiDEA_ERP_Server/SECUiDEA_ERP_Server/Controllers/BaseControllers/JwtController.cs
@@ -25,7 +25,7 @@ public class JwtController : BaseController
         }
         else
         {
-            int days = rememberMe
+            double days = rememberMe
                 ? _jwtService.GetSettings.AutoLoginDays
                 : _jwtService.GetSettings.RefreshTokenDays;
